Homogenize file size error from nginx with python errors
JIRA: DOVETAIL-732

This patch homogenize error received from nginx when a file is bigger
than the limit set with python errors. It tells nginx to convert
error 413 to 200 and respond with a json, like python does when
validating the file. As a result, the error is shown in the pop up
as it should be.

Change-Id: I3218516073f848321af02465c5d608ef54d5b715
Signed-off-by: Stamatis Katsaounis <mokats@intracom-telecom.com>