Enhanced OVP Web Portal - minor improvements
The OVP web portal is updated with some minor improvements.
Those improvements are focused on the styling part of the portal
as well as on the introduction of a new tab for upcoming reviews
(giving the opportunity to the reviewers to see their-own results
separately with the results which waiting for review).

JIRA: DOVETAIL-760

Change-Id: Ic5ca90eaaee36de43486ff55ad9aa788136587c6
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>