Update OVP portal, Add skipped TCs in the report
Since the dovetail supports the 'SKIP' verdict for the results,
the OVP portal should be updated in order to collect and appear the
skipped testcases and sub-testcases properly.

The report of results has been improved, so that a third column is
introduced for skipped results (Passed, Skipped, Not Passed).

JIRA: DOVETAIL-765

Change-Id: Ia1a0d8bd732814137e495dc4c6ef5ebb0ef0ed73
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>