Update patch set 2
Patch Set 2:

(3 comments)

I added some comments for improvements but I would also like to ask the whole team:

If we decide that we want the SKIP status (I am positive) then we have to ask for TSC approval? In addition, we are going to introduce it to the current scope as well?

Patch-set: 2
Label: Verified=0