Update patch set 2
Patch Set 2:

(1 comment)

I am quite confused about this patch. In the commit message, it said "updates for NFVI Launch", but everything it changed is for ONAP UI which is "VNF".
Also there are OVP somewhere in ONAP UI. To my understanding, OVP is only for OPNFV i.e. NFVI. Should not make OVP cover both NFVI and VNF. Maybe change it to be CVC will be better.

Patch-set: 2
Label: Verified=0