Update the behavior of roles for OVP/ONAP portals
The behavior of portals' roles are updated according to CVC decisions.
The behavior will be:
* no roles - the person can upload results and share results
to other users, but can not submit an application/results for review
* "user" role - the person can submit the application/results
fore review. To have this role, the person MUST have signed the
participation agreement through docusign.
* "reviewer" role - the person can see applications / results
submitted to review, and can vote (+1 / -1) on those results as part
of the review process. Votes are recorded.
* "admin" role - the person can approve applications
for public listing, etc. Final gatekeeper for the website, etc.
Responsible to ensure processes are followed, etc.

Change-Id: Ieca55725378c85f544eeea4390046bae456ee1a3
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>