Update patch set 1
Patch Set 1:

(7 comments)

I guess many problems are caused by the backend API. when requesting the APIs, it couldn't get the right data.
Don't suggest to filter the data with the frontend which is not security. All data have actually got from DB but not shown on the page.

Patch-set: 1