Update patch set 2
Patch Set 2:

(2 comments)

I have test it locally and all the problems we found before are seems to be fixed.
I have come across a problem quite occasionally twice but can't reproduce it again. The problem is that after approving one result (showing it in the homepage and its status is 'verified'), doing some operations, then the application disappeared from both the DB and the homepage, but the status is still  'verified'.

Patch-set: 2
Label: Verified=0