Portal Bug - non proper display of Non Exec TCs
The non executed test cases have not any verdict,
assigned by the dovetail project (e.g. PASS, SKIP, FAIL).
In that case, the portal cannot recognize those TCs as failed ones
and they are out of the 'Not Passed' list in the Test Result Overview.

Therefore, the controller has been updated, removing the check for the failed TCs.
If the TC is not belonged either to Pass or Skip list then this TC is pushed
in the Not Pass list. So, the failed as well as the non executed TCs appear
under 'Non Passed' list

JIRA: DOVETAIL-773

Change-Id: I71e639c168bbda1065e12b063e22f17875536a9a
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>